Refactoring tests to proper .spec and resolve conversation due to comments.
