+++ src/test/groovy/org/prebid/server/functional/tests/AmpSpec.groovy
@@ -16,116 +16,6 @@ import static org.prebid.server.functional.util.SystemProperties.PBS_VERSION
 
 class AmpSpec extends BaseSpec {
 
-    private static final int DEFAULT_TIMEOUT = getRandomTimeout()
-    private static final String PBS_VERSION_HEADER = "pbs-java/$PBS_VERSION"
-
-    @Shared
-    PrebidServerService prebidServerService = pbsServiceFactory.getService(["auction.max-timeout-ms"    : MAX_TIMEOUT as String,
-                                                                            "auction.default-timeout-ms": DEFAULT_TIMEOUT as String])
-
-    def "PBS should apply timeout from stored request when it's not specified in the request"() {
-        given: "Default AMP request without timeout"
-        def ampRequest = AmpRequest.defaultAmpRequest.tap {
-            timeout = null
-        }
-
-        and: "Default stored request with timeout"
-        def timeout = getRandomTimeout()
-        def ampStoredRequest = BidRequest.defaultStoredRequest.tap {
-            tmax = timeout
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        prebidServerService.sendAmpRequest(ampRequest)
-
-        then: "Bidder request should contain timeout from the stored request"
-        def bidderRequest = bidder.getBidderRequest(ampStoredRequest.id)
-        assert bidderRequest.tmax == timeout as Long
-    }
-
-    def "PBS should prefer timeout from the request when stored request timeout is #tmax"() {
-        given: "Default AMP request with timeout"
-        def timeout = getRandomTimeout()
-        def ampRequest = AmpRequest.defaultAmpRequest.tap {
-            it.timeout = timeout
-        }
-
-        and: "Default stored request"
-        def ampStoredRequest = BidRequest.defaultStoredRequest.tap {
-            it.tmax = tmax
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        prebidServerService.sendAmpRequest(ampRequest)
-
-        then: "Bidder request should contain timeout from the request"
-        def bidderRequest = bidder.getBidderRequest(ampStoredRequest.id)
-        assert bidderRequest.tmax == timeout as Long
-
-        where:
-        tmax << [null, getRandomTimeout()]
-    }
-
-    def "PBS should honor max timeout from the settings"() {
-        given: "Default AMP request"
-        def ampRequest = AmpRequest.defaultAmpRequest.tap {
-            it.timeout = ampRequestTimeout
-        }
-
-        and: "Default stored request"
-        def ampStoredRequest = BidRequest.defaultStoredRequest.tap {
-            it.tmax = storedRequestTimeout
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        prebidServerService.sendAmpRequest(ampRequest)
-
-        then: "Bidder request timeout should correspond to the maximum from the settings"
-        def bidderRequest = bidder.getBidderRequest(ampStoredRequest.id)
-        assert bidderRequest.tmax == MAX_TIMEOUT as Long
-
-        where:
-        ampRequestTimeout || storedRequestTimeout
-        MAX_TIMEOUT + 1   || null
-        null              || MAX_TIMEOUT + 1
-        MAX_TIMEOUT + 1   || MAX_TIMEOUT + 1
-    }
-
-    def "PBS should honor default timeout"() {
-        given: "Default AMP request without timeout"
-        def ampRequest = AmpRequest.defaultAmpRequest.tap {
-            it.timeout = null
-        }
-
-        and: "Default stored request without timeout"
-        def ampStoredRequest = BidRequest.defaultStoredRequest.tap {
-            it.tmax = null
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        prebidServerService.sendAmpRequest(ampRequest)
-
-        then: "Bidder request timeout should correspond to the maximum from the settings"
-        def bidderRequest = bidder.getBidderRequest(ampStoredRequest.id)
-        assert bidderRequest.tmax == DEFAULT_TIMEOUT as Long
-    }
-
     def "PBS should return version in response header for #description"() {
         given: "Default AmpRequest"
         def ampStoredRequest = BidRequest.defaultBidRequest
@@ -139,7 +29,7 @@ class AmpSpec extends BaseSpec {
         def response = defaultPbsService.sendAmpRequestRaw(ampRequest)
 
         then: "Response header should contain PBS version"
-        assert response.headers["x-prebid"] == PBS_VERSION_HEADER
+        assert response.headers["x-prebid"] == "pbs-java/$PBS_VERSION"
 
         where:
         ampRequest                   || description
