+++ src/test/groovy/org/prebid/server/functional/tests/AuctionSpec.groovy
@@ -5,88 +5,15 @@ import org.prebid.server.functional.model.db.Account
 import org.prebid.server.functional.model.db.StoredRequest
 import org.prebid.server.functional.model.request.auction.BidRequest
 import org.prebid.server.functional.model.request.auction.PrebidStoredRequest
-import org.prebid.server.functional.model.response.auction.ErrorType
 import org.prebid.server.functional.service.PrebidServerException
-import org.prebid.server.functional.service.PrebidServerService
-import org.prebid.server.functional.testcontainers.container.PrebidServerContainer
 import org.prebid.server.functional.util.PBSUtils
-import org.testcontainers.utility.MountableFile
-import spock.lang.Shared
 
 import static org.prebid.server.functional.model.AccountStatus.INACTIVE
 import static org.prebid.server.functional.model.request.auction.DistributionChannel.APP
 import static org.prebid.server.functional.model.request.auction.DistributionChannel.SITE
-import static org.prebid.server.functional.testcontainers.container.PrebidServerContainer.APP_WORKDIR
 import static org.prebid.server.functional.util.SystemProperties.PBS_VERSION
 
 class AuctionSpec extends BaseSpec {
-
-    private static final int DEFAULT_TIMEOUT = getRandomTimeout()
-    private static final String PBS_VERSION_HEADER = "pbs-java/$PBS_VERSION"
-
-    @Shared
-    PrebidServerService prebidServerService = pbsServiceFactory.getService(["auction.max-timeout-ms"    : MAX_TIMEOUT as String,
-                                                                            "auction.default-timeout-ms": DEFAULT_TIMEOUT as String])
-
-    def "PBS should apply timeout from stored request when it's not specified in the auction request and pbs in config is lowest"() {
-        given: "PBs config with default-timeout-ms"
-        def prebidServerService = pbsServiceFactory.getService(["auction.default-timeout-ms": "10"])
-
-        and: "Default BidRequest without timeout"
-        def bidRequest = BidRequest.defaultBidRequest.tap {
-            tmax = null
-            ext.prebid.storedRequest = new PrebidStoredRequest(id: PBSUtils.randomNumber)
-        }
-
-        and: "Default stored request with minTimeout"
-        def timeout = 1000L
-        def storedRequest = BidRequest.defaultStoredRequest.tap {
-            tmax = timeout
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequestModel = StoredRequest.getStoredRequest(bidRequest, storedRequest)
-        storedRequestDao.save(storedRequestModel)
-
-        when: "PBS processes auction request"
-        prebidServerService.sendAuctionRequest(bidRequest)
-
-        then: "Bidder request should contain timeout from the stored request"
-        def bidderRequest = bidder.getBidderRequest(bidRequest.id)
-        assert bidderRequest.tmax == timeout as Long
-    }
-
-    def "PBS should return error when timeout from stored request is lowest"() {
-        given: "PBs config with default-timeout-ms"
-        def prebidServerService = pbsServiceFactory.getService(["auction.default-timeout-ms": "1000"])
-
-        and: "Default BidRequest without timeout"
-        def bidRequest = BidRequest.defaultBidRequest.tap {
-            tmax = null
-            ext.prebid.storedRequest = new PrebidStoredRequest(id: PBSUtils.randomNumber)
-        }
-
-        and: "Default stored request with minTimeout"
-        def minTmax = 1L
-        def storedRequest = BidRequest.defaultStoredRequest.tap {
-            tmax = minTmax
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequestModel = StoredRequest.getStoredRequest(bidRequest, storedRequest)
-        storedRequestDao.save(storedRequestModel)
-
-        when: "PBS processes auction request"
-        def bidResponse = prebidServerService.sendAuctionRequest(bidRequest)
-
-        then: "Bidder request timeout should correspond to the min from the stored request"
-        assert bidResponse.ext.debug.resolvedRequest.tmax == minTmax
-
-        and: "Bid response should shutdown by timeout from stored request"
-        def errors = bidResponse?.ext?.errors
-        assert errors[ErrorType.GENERIC]*.code == [1]
-        assert errors[ErrorType.GENERIC]*.message == ["Timeout has been exceeded"]
-    }
 
     def "PBS should return version in response header for auction request for #description"() {
 
@@ -94,169 +21,12 @@ class AuctionSpec extends BaseSpec {
         def response = defaultPbsService.sendAuctionRequestRaw(bidRequest)
 
         then: "Response header should contain PBS version"
-        assert response.headers["x-prebid"] == PBS_VERSION_HEADER
+        assert response.headers["x-prebid"] == "pbs-java/$PBS_VERSION"
 
         where:
         bidRequest                   || description
         BidRequest.defaultBidRequest || "valid bid request"
         new BidRequest()             || "invalid bid request"
-    }
-
-    def "PBS should apply timeout from stored request when it's not specified in the auction request"() {
-        given: "Default basic BidRequest with generic bidder"
-        def bidRequest = BidRequest.defaultBidRequest.tap {
-            tmax = null
-            ext.prebid.storedRequest = new PrebidStoredRequest(id: PBSUtils.randomNumber)
-        }
-
-        and: "Default stored request with timeout"
-        def timeout = getRandomTimeout()
-        def storedRequestModel = BidRequest.defaultStoredRequest.tap {
-            tmax = timeout
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getStoredRequest(bidRequest, storedRequestModel)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes auction request"
-        prebidServerService.sendAuctionRequest(bidRequest)
-
-        then: "Bidder request should contain timeout from the stored request"
-        def bidderRequest = bidder.getBidderRequest(bidRequest.id)
-        assert bidderRequest.tmax == timeout as Long
-    }
-
-    def "PBS should prefer timeout from the auction request"() {
-        given: "Default basic BidRequest with generic bidder"
-        def timeout = getRandomTimeout()
-        def bidRequest = BidRequest.defaultBidRequest.tap {
-            tmax = timeout
-            ext.prebid.storedRequest = new PrebidStoredRequest(id: PBSUtils.randomNumber)
-        }
-
-        and: "Default stored request"
-        def storedRequestModel = BidRequest.defaultStoredRequest.tap {
-            it.tmax = tmaxStoredRequest
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getStoredRequest(bidRequest, storedRequestModel)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes auction request"
-        prebidServerService.sendAuctionRequest(bidRequest)
-
-        then: "Bidder request should contain timeout from the request"
-        def bidderRequest = bidder.getBidderRequest(bidRequest.id)
-        assert bidderRequest.tmax == timeout as Long
-
-        where:
-        tmaxStoredRequest << [null, getRandomTimeout()]
-    }
-
-    def "PBS should honor max timeout from the settings for auction request"() {
-        given: "Default basic BidRequest with generic bidder"
-        def bidRequest = BidRequest.defaultBidRequest.tap {
-            tmax = auctionRequestTimeout
-            ext.prebid.storedRequest = new PrebidStoredRequest(id: PBSUtils.randomNumber)
-        }
-
-        and: "Default stored request"
-        def storedRequest = BidRequest.defaultStoredRequest.tap {
-            it.tmax = storedRequestTimeout
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequestModel = StoredRequest.getStoredRequest(bidRequest, storedRequest)
-        storedRequestDao.save(storedRequestModel)
-
-        when: "PBS processes auction request"
-        prebidServerService.sendAuctionRequest(bidRequest)
-
-        then: "Bidder request timeout should correspond to the maximum from the settings"
-        def bidderRequest = bidder.getBidderRequest(bidRequest.id)
-        assert bidderRequest.tmax == MAX_TIMEOUT as Long
-
-        where:
-        auctionRequestTimeout || storedRequestTimeout
-        MAX_TIMEOUT + 1       || null
-        null                  || MAX_TIMEOUT + 1
-        MAX_TIMEOUT + 1       || MAX_TIMEOUT + 1
-    }
-
-    def "PBS should honor default timeout for auction request"() {
-        given: "Default basic BidRequest without timeout"
-        def bidRequest = BidRequest.defaultBidRequest.tap {
-            tmax = null
-            ext.prebid.storedRequest = new PrebidStoredRequest(id: PBSUtils.randomNumber)
-        }
-
-        and: "Default stored request without timeout"
-        def storedRequest = BidRequest.defaultStoredRequest.tap {
-            it.tmax = null
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequestModel = StoredRequest.getStoredRequest(bidRequest, storedRequest)
-        storedRequestDao.save(storedRequestModel)
-
-        when: "PBS processes auction request"
-        prebidServerService.sendAuctionRequest(bidRequest)
-
-        then: "Bidder request timeout should correspond to the maximum from the settings"
-        def bidderRequest = bidder.getBidderRequest(bidRequest.id)
-        assert bidderRequest.tmax == DEFAULT_TIMEOUT as Long
-    }
-
-    def "PBS should take data by priority when request, stored request, default request are defined"() {
-        given: "Default request with timeout"
-        def defaultRequestModel = new BidRequest(tmax: defaultRequestTmax)
-        def defaultRequest = PBSUtils.createJsonFile(defaultRequestModel)
-
-        and: "Pbs config with default request"
-        def pbsContainer = new PrebidServerContainer(
-                ["default-request.file.path" : APP_WORKDIR + defaultRequest.fileName,
-                 "auction.max-timeout-ms"    : MAX_TIMEOUT as String,
-                 "auction.default-timeout-ms": DEFAULT_TIMEOUT as String]).tap {
-            withCopyFileToContainer(MountableFile.forHostPath(defaultRequest), APP_WORKDIR) }
-        pbsContainer.start()
-        def pbsService = new PrebidServerService(pbsContainer)
-
-        and: "Default basic BidRequest with timeout"
-        def bidRequest = BidRequest.defaultBidRequest.tap {
-            tmax = requestTmax
-            ext.prebid.storedRequest = new PrebidStoredRequest(id: PBSUtils.randomNumber)
-        }
-
-        and: "Default stored request with timeout"
-        def storedRequestModel = BidRequest.defaultStoredRequest.tap {
-            tmax = storedRequestTmax as Long
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getStoredRequest(bidRequest, storedRequestModel)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes auction request"
-        def response = pbsService.sendAuctionRequest(bidRequest)
-
-        then: "Response should not contain error"
-        assert !response.ext?.errors
-        assert !response.ext?.warnings
-
-        then: "Bidder request should contain correct tmax"
-        def bidderRequest = bidder.getBidderRequest(bidRequest.id)
-        assert bidderRequest.tmax == DEFAULT_TIMEOUT as Long
-
-        cleanup: "Stop container with default request"
-        pbsContainer.stop()
-
-        where:
-        requestTmax     | storedRequestTmax  | defaultRequestTmax
-        DEFAULT_TIMEOUT | getRandomTimeout() | getRandomTimeout()
-        null            | DEFAULT_TIMEOUT    | getRandomTimeout()
-        null            | null               | DEFAULT_TIMEOUT
     }
 
     def "PBS should update account.<account-id>.requests.rejected.invalid-account metric when account is inactive"() {
@@ -365,7 +135,7 @@ class AuctionSpec extends BaseSpec {
         storedRequestDao.save(storedRequest)
 
         when: "Requesting PBS auction"
-        prebidServerService.sendAuctionRequest(bidRequest)
+        defaultPbsService.sendAuctionRequest(bidRequest)
 
         then: "BidResponse should be merged with stored request"
         def bidderRequest = bidder.getBidderRequest(bidRequest.id)
@@ -384,7 +154,7 @@ class AuctionSpec extends BaseSpec {
         }
 
         when: "Requesting PBS auction"
-        def response = prebidServerService.sendAuctionRequest(bidRequest)
+        def response = defaultPbsService.sendAuctionRequest(bidRequest)
 
         then: "BidResponse should contain the same passThrough as on request"
         assert response.seatbid.first().bid.first().ext.prebid.passThrough == passThrough
@@ -399,7 +169,7 @@ class AuctionSpec extends BaseSpec {
         }
 
         when: "Requesting PBS auction"
-        prebidServerService.sendAuctionRequest(bidRequest)
+        defaultPbsService.sendAuctionRequest(bidRequest)
 
         then: "BidResponse should contain the same passThrough as on request"
         def bidderRequest = bidder.getBidderRequest(bidRequest.id)
